Yarn v2 (#4)
Move to Yarn v2 for library and demo/, which makes it possible to use
the `portal:` protocol and increases installation speed (even though PnP
is not enabled for now).

The demo now deduplicates react and react-dom, which are required by the
library as devDendencies (for Jest) and was requiring to copy the
library in demo/node_modules, then manually delete
demo/node_modules/use-nft/node_modules/react. This shouldn’t be needed
anymore and has been removed from the gh-pages GitHub Action.